// types/inventoryVoucher/SalesVoucherTypes.d.ts
interface InvoiceModalProps {
  showInvoiceDialog: boolean;
  setShowInvoiceDialog: Dispatch<SetStateAction<boolean>>;
}
// ...
export interface InvoiceTableData {
  Id: number;
  Sale_Date: string;
  Sale_No: string;
  Party_Name: string;
  Party_Id: number;
  Party_Add: string;
  Party_Mob: string;
  Party_GST: string;
  Tot_Amount: string;
  CGST_Rate: number;
  Tot_CGST: string;
  SGST_Rate: number;
  Tot_SGST: string;
  Tot_IGST: string;
  Tot_Round: string;
  Tot_Disc: string;
  DesignRow: {
    Design_Id: number;
    Design_Name: string;
    Design_No: string;
    Order_Qnty: string;
    Design_Rate: string;
    Wt: string;
    Wt_Rate: string;
    Tot_Wt: string;
    Polish: string;
    Tot_Polish: string;
    Image: string;
    ItemRow: {
      Item_Id: number;
      Item_Name: string;
      Item_Qnty: string;
      Item_Rate: string;
      Making_Rate: string;
      Item_Tot: string;
    }[];
  }[];
}